+++ CHANGELOG.md
@@ -51,6 +51,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `opentelemetry-instrumentation-aiohttp-client` aiohttp: Correct url filter input type
   ([#843](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/864))
 
+- `opentelemetry-instrumentation-aiohttp-client` aiohttp: Remove `span_name` from docs
+  ([#857](https://github.com/open-telemetry/opentelemetry-python-contrib/issues/857))
+
 
 ## [1.8.0-0.27b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.8.0-0.27b0) - 2021-12-17